+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 - New readers for export improve memory usage loading small batches of products instead of all products in same time
 - Update BatchBundle to 0.1.6 in order to get updated summary information during the execution of the process (and not only at the end)
 - Allow values 'true', 'false', 'yes' and 'no' to be converted into boolean during import
+- Create a job instance factory to create job instances
+- Allow to add hidden row actions in grids
+- Make optional the generation of missing completenesses in product reader
 
 ## Bug fixes
 - Replaced usage of Symfony process to launch background job with a simple exec, more reliable on a heavily loaded environment
@@ -72,6 +75,7 @@
 - Remove WidgetRepository to replace it by `Pim\Bundle\ImportExportBundle\Entity\Repository\JobExecutionRepository`
 - Inject `Pim\Bundle\ImportExportBundle\Manager\JobExecutionManager` into LastOperationsWidget
 - Remove injection of WidgetRepository from LastOperationsWidget
+- Inject JobInstanceFactory inside `Pim\Bundle\ImportExportBundle\Controller\JobProfileController`
 
 # 1.1.0 - "Rabbit Punch" (2014-04-16)